rename burnside namespace to lower
