+++ lib/CMakeLists.txt
@@ -12,10 +12,10 @@ if ((CMAKE_CXX_COMPILER_ID MATCHES "Clang"))
   endif()
 endif()
 
-add_subdirectory(burnside)
 add_subdirectory(common)
 add_subdirectory(decimal)
 add_subdirectory(evaluate)
 add_subdirectory(fir)
+add_subdirectory(lower)
 add_subdirectory(parser)
 add_subdirectory(semantics)
