+++ lib/lower/CMakeLists.txt
@@ -14,7 +14,7 @@
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-error")
 
-add_library(FortranBurnside
+add_library(FortranLower
   ast-builder.cc
   bridge.cc
   builder.cc
@@ -26,14 +26,14 @@ add_library(FortranBurnside
   io.cc
 )
 
-target_link_libraries(FortranBurnside
+target_link_libraries(FortranLower
   FIR
   MLIRAffineOps
   MLIRLLVMIR
   MLIRStandardOps
 )
 
-install (TARGETS FortranBurnside
+install (TARGETS FortranLower
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
   RUNTIME DESTINATION bin
