+++ lib/lower/bridge.cc
@@ -1,4 +1,4 @@
-//===-- lib/burnside/bridge.cc ----------------------------------*- C++ -*-===//
+//===-- lib/lower/bridge.cc -------------------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -28,7 +28,7 @@
 #include "mlir/Parser.h"
 #include "mlir/Target/LLVMIR.h"
 
-namespace Br = Fortran::burnside;
+namespace Br = Fortran::lower;
 namespace Co = Fortran::common;
 namespace Ev = Fortran::evaluate;
 namespace L = llvm;
@@ -37,7 +37,7 @@ namespace Pa = Fortran::parser;
 namespace Se = Fortran::semantics;
 
 using namespace Fortran;
-using namespace Fortran::burnside;
+using namespace Fortran::lower;
 
 namespace {
